Base text track implementation for video element bug 67933
git-svn-id: https://svn.intranet.ydp/ydp/projects/empiria/player/trunk/empiria.player@114788 3e83ccf1-c6ce-0310-9054-8fdb33cf0640

+++ src/eu/ydp/empiria/player/client/controller/extensions/internal/sound/AbstractMediaProcessor.java
@@ -41,7 +41,7 @@ public abstract class AbstractMediaProcessor extends InternalExtension implement
 
 		switch (((MediaEventTypes) event.getAssociatedType().getType())) {
 		case CHANGE_VOLUME:
-			executor.setVolume(event.getValue());
+			executor.setVolume((Double) event.getValue());
 			break;
 		case STOP:
 			executor.stop();
@@ -50,7 +50,7 @@ public abstract class AbstractMediaProcessor extends InternalExtension implement
 			executor.pause();
 			break;
 		case SET_CURRENT_TIME:
-			executor.setCurrentTime(event.getValue());
+			executor.setCurrentTime((Double) event.getValue());
 			break;
 		case PLAY:
 			executor.play();
